add handling for loop one song (e.g. for the ghostbusters command)
